+++ submarine/__init__.py
@@ -1,4 +1,4 @@
-from submarine.tools import (
+from submarine.systems import (
     Radar,
     Navigation,
 )
